Refactoring project structure & build script.
