+++ bower.json
@@ -1,19 +1,10 @@
 {
   "name": "dosomething-validation",
-  "version": "0.1.0",
   "main": [
     "dist/validation.css",
     "dist/validation.js"
   ],
   "dependencies": {
     "jquery": "^1.7"
-  },
-  "devDependencies": {
-    "jquery": "1.8.3",
-    "almond": "~0.2.9",
-    "neue": "DoSomething/neue#^6.0.0"
-  },
-  "resolutions": {
-    "jquery": "1.8.3"
   }
 }
