--- tasks/aliases.json
@@ -1,5 +1,0 @@
-{
-  "default": ["build", "test", "watch"],
-  "build": ["clean:dist", "sass", "postcss", "requirejs", "modernizr"],
-  "test": ["jshint", "qunit"]
-}
