--- tasks/bump.js
@@ -1,8 +1,0 @@
-module.exports = {
-  "options": {
-    "files": ["package.json", "bower.json"],
-    "commitFiles": ["package.json", "bower.json"],
-    "push": false,
-    "createTag": false
-  }
-};
