--- tasks/qunit.js
@@ -1,3 +1,0 @@
-module.exports = {
-  all: ["tests/*.html"]
-};
