+++ tests/index.html
@@ -1,39 +1,44 @@
 <!doctype html>
 <html>
-  <head>
-      <meta charset='UTF-8' />
-      <meta http-equiv='content-type' content='text/html; charset=utf-8' />
+<head>
+  <meta charset='UTF-8' />
+  <meta http-equiv='content-type' content='text/html; charset=utf-8' />
 
-      <title>Modal QUnit Test Runner</title>
+  <title>Modal QUnit Test Runner</title>
 
-      <script src='../bower_components/jquery/jquery.js'></script>
-      <link rel='stylesheet' href='../node_modules/grunt-contrib-qunit/test/libs/qunit.css'>
-      <script src='../node_modules/grunt-contrib-qunit/test/libs/qunit.js'></script>
+  <link rel='stylesheet' href='../node_modules/qunitjs/qunit/qunit.css'>
+  <script src='../node_modules/qunitjs/qunit/qunit.js'></script>
 
-      <!-- add any external libraries your code needs -->
+  <script>
+    QUnit.config.autostart = false;
+  </script>
 
-      <!-- add any JS files under test (or put them in different .html files) -->
-      <link rel="stylesheet" href="../dist/modal.css"/>
-      <script src='../dist/modal.js'></script>
+  <!-- add any external libraries your code needs -->
+  <script src="../node_modules/jquery/dist/jquery.js"></script>
+  <script src='../dist/modernizr.js'></script>
 
-      <!-- your tests, any and all to run with the given fixtures below -->
-      <script src='tests.js'></script>
+  <!-- add any JS files under test (or put them in different .html files) -->
+  <link rel="stylesheet" href="../dist/validation.css"/>
+  <script src='../dist/validation.js'></script>
 
-  </head>
-  <body>
-    <div class="chrome">
+  <!-- your tests, any and all to run with the given fixtures below -->
+  <script src='tests.js'></script>
 
-      <a href="#" data-modal-href="#test-modal">Test Link</a>
+</head>
+<body>
+<div class="chrome">
 
-      <div data-modal id="test-modal">
-        <p>It's working!</p>
-      </div>
+  <div id="qunit"></div> <!-- QUnit fills this with results, etc -->
+  <div id='qunit-fixture'>
+    <!-- any HTML you want to be present in each test (will be reset for each test) -->
+  </div>
 
-      <div id="qunit"></div> <!-- QUnit fills this with results, etc -->
-      <div id='qunit-fixture'>
-        <!-- any HTML you want to be present in each test (will be reset for each test) -->
-      </div>
+  <script>
+    $(document).ready(function() {
+      start();
+    });
+  </script>
 
-    </div>
-  </body>
+</div>
+</body>
 </html>
